backend/requirements.txt
# Gerado a partir do pyproject.toml - use: pip install -r requirements.txt
# Ou instale o projeto em modo editável: pip install -e .

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
cachetools>=5.3.2
openpyxl>=3.1.2
reportlab>=4.0.8
structlog>=24.1.0
